+++ src/components/Products/CustomerCard/index.tsx
@@ -1,22 +1,38 @@
 import React from 'react'
 import { CallToAction } from '../../CallToAction'
+import { layoutLogic } from 'logic/layoutLogic'
+import { useValues } from 'kea'
 
 interface CustomerCardProps {
-    logo: string
     outcome: string
     quote: string
-    link: string
+    customer: {
+        fields: {
+            slug: string
+        }
+        frontmatter: {
+            logo: {
+                publicURL: string
+            }
+            logoDark: {
+                publicURL: string
+            }
+        }
+    }
 }
 
-export const CustomerCard = ({ logo, outcome, quote, link }: CustomerCardProps): JSX.Element => {
+export const CustomerCard = ({ outcome, quote, customer }: CustomerCardProps): JSX.Element => {
+    const { websiteTheme } = useValues(layoutLogic)
+    const darkMode = websiteTheme === 'dark'
+
     return (
         <li className=" bg-accent dark:bg-accent-dark p-4 rounded">
-            <div>{logo}</div>
+            <img className="mb-4 max-w-[200px]" src={customer.frontmatter[darkMode ? 'logoDark' : 'logo'].publicURL} />
             <p className="text-lg font-semibold m-0 leading-tight mb-1">{outcome}</p>
             <p className="text-sm">
                 <span className="bg-highlight p-0.5">"{quote}"</span>
             </p>
-            <CallToAction href={link} type="secondary" size="sm">
+            <CallToAction to={customer.fields.slug} type="secondary" size="sm">
                 Read the story
             </CallToAction>
         </li>
