+++ src/components/Products/Feature/index.tsx
@@ -4,14 +4,14 @@ import { StaticImage } from 'gatsby-plugin-image'
 interface FeatureProps {
     name: string
     description: string
-    image: string
+    image: any
 }
 
-export const Feature = ({ name, description }: FeatureProps): JSX.Element => {
+export const Feature = ({ name, description, image }: FeatureProps): JSX.Element => {
     return (
         <li className="text-center">
-            <div className="mb-2">
-                <StaticImage src="../images/products/session-replay-timeline.png" className="max-w-[421px]" />
+            <div className="mb-2 w-full relative after:absolute after:w-full after:h-24 after:bottom-0 after:left-0 after:bg-gradient-to-b after:from-transparent dark:after:via-dark/80 dark:after:to-dark after:via-light/80 after:to-light after:z-10">
+                {image}
             </div>
             <h4 className="mb-0 text-lg">{name}</h4>
             <p className="text-[15px]">{description}</p>
